refactor: separate view and logic in members component
