+++ src/screens/MembersScreen/index.tsx
@@ -1,4 +1,4 @@
-import Members from '@/features/members/components';
+import Members from '@/features/members/components/Members/container';
 
 const MembersScreen = () => {
   return <Members />;
